wikipedia_pageviews_coresentiment_pipeline/requirements.txt
requests
apache-airflow-providers-postgres
psycopg2-binary
python-dotenv
sqlalchemy